Move the client tests with the client they cover
Four test files stayed in public-api-client importing ../../src/index.js,
which no longer exports the hand-written operations - they moved to
public-api-react with the source. Verified they run there rather than
trusting the file count: intake.test.ts alone executes 4 assertions.
