+++ packages/public-api-react/tests/unit/legacy-client/booking-session-v1.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest"
 
-import { createVoyantPublicApiClient } from "../../src/index.js"
+import { createVoyantPublicApiClient } from "../../../src/legacy-client/index.js"
 
 const REQUIREMENTS = {
   showsConfigure: true,
